Create ZZPR62FL.DTA
This is the PR model dataset, it is also used along with the IR model dataset ZZIR62FL.dta to demonstrate the use of the code in this repository. This is not real data from any country.

Read more about DHS Program model datasets here: https://www.dhsprogram.com/data/Model-Datasets.cfm